Move component files closer to where they are used and make function docs consistent
